# openapi-spec.yaml
openapi: "3.0.3"
info:
  title: "Fluxi API - Multichannel Inventory Management"
  version: "1.0.0"
  description: |
    Complete API for Fluxi - a multichannel inventory management system with ERP integrations,
    e-commerce platforms, and centralized stock control. Supports multi-tenancy, real-time sync,
    and automated order fulfillment.

    ## Key Features:
    - **Multi-tenant architecture** with account isolation
    - **Plugin-based channel integrations** (Siigo, Shopify, WooCommerce, etc.)
    - **Centralized inventory management** with multiple warehouses
    - **Product catalog (PIM)** with variants and mappings
    - **Automated order processing** and fulfillment
    - **Real-time synchronization** with external platforms
    - **Comprehensive audit logging** and reporting
servers:
  - url: https://api.fluxi.com/v1
    description: Production API
  - url: https://staging.api.fluxi.com/v1
    description: Staging API
  - url: http://localhost:4000
    description: Local development
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string
        details:
          type: object
      required:
        - success
        - error
...
    UserCredentials:
      type: object
      properties:
        email:
          type: string
          format: email
          example: "user@company.com"
        password:
          type: string
          minLength: 8
          example: "securePassword123"
      required:
        - email
        - password
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        user:
          $ref: '#/components/schemas/User'
        token:
          type: string
          description: JWT token for authentication
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      required:
        - success
        - user
        - token
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        full_name:
          type: string
        created_at:
          type: string
          format: date-time
        last_login_at:
          type: string
          format: date-time

    # Account schemas
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Mi Empresa S.A.S."
        slug:
          type: string
          example: "mi-empresa"
        default_currency:
          type: string
          example: "COP"
        timezone:
          type: string
          example: "America/Bogota"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
    AccountUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        account_id:
          type: string
          format: uuid
        role:
          type: string
          enum: [owner, admin, manager, member]
        is_owner:
          type: boolean
        created_at:
          type: string
          format: date-time

    # Channel schemas
...
paths:
...
  /auth/login:
    post:
      summary: User login
      description: Authenticate user and return JWT token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UserCredentials' }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # =========================================
  # ACCOUNTS MANAGEMENT
  # =========================================
  /accounts:
    get:
      summary: List user accounts
      description: Get all accounts where the authenticated user has access
      operationId: listAccounts
      security: [ BearerAuth: [] ]
      responses:
        '200':
          description: Accounts retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  accounts:
                    type: array
                    items: { $ref: '#/components/schemas/Account' }
        '401':
          description: Unauthorized
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

    post:
      summary: Create new account
      description: Create a new account for the authenticated user (as owner)
      operationId: createAccount
      security: [ BearerAuth: [] ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  example: "Nueva Empresa S.A.S."
                default_currency:
                  type: string
                  example: "COP"
                timezone:
                  type: string
                  example: "America/Bogota"
              required:
                - name
      responses:
        '201':
          description: Account created successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  account:
                    $ref: '#/components/schemas/Account'
        '400':
          description: Invalid input
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /accounts/{accountId}/users:
    get:
      summary: List account users
      description: Get all users with access to the account
      operationId: listAccountUsers
      security: [ BearerAuth: [] ]
      parameters:
        - name: accountId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  users:
                    type: array
                    items: { $ref: '#/components/schemas/AccountUser' }
        '403':
          description: No access to account
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

    post:
      summary: Invite user to account
      description: Invite a new user to join the account with specified role
      operationId: inviteAccountUser
      security: [ BearerAuth: [] ]
      parameters:
        - name: accountId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                role:
                  type: string
                  enum: [admin, manager, member]
                  default: member
              required:
                - email
      responses:
        '201':
          description: User invited successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Invitation sent to user@email.com"
        '400':
          description: Invalid input or user already exists
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }

  # =========================================
  # CHANNEL MANAGEMENT
  # =========================================
...